Added MinimalTryCatch referee.
